Enable the Ledger and drop outdated packages
A CosmosApp class was derived from IovLedgerApp and used as an adapter as a drop-in for the existing work Forble did.

NOTE: Apparently there was a bug in lunie-cosmos-js that necessitated createCosmosAddress(); however, it's not needed anymore.

BTW: https://stackoverflow.com/a/35092034/5887599 shows a nice trick to secure an http connection, which is required for
TransportWebUSB:

Or you could just use ngrok to port forward :)

1) start your server (i.e. at localhost:3000)

2) start ngrok from command line: ./ngrok http 3000

that should give you http and https urls to access from any device